Moved all non route and non essential related file to outside the src folder
