+++ src/app/dashboard/loading.tsx
@@ -1,8 +1,7 @@
 "use client";
-import { Box, Container, Divider } from "@mui/material";
-import Skeleton from "@mui/material/Skeleton";
-import DashboardHeader from "../components/dashboard/dashboardHeader";
-import DashBoardSkeleton from "../components/skeletons/dashboardSkeleton";
+
+import DashBoardSkeleton from "@/components/skeletons/dashboardSkeleton";
+
 
 export default function Loading() {
   return <DashBoardSkeleton />;
