+++ src/app/page.tsx
@@ -1,9 +1,10 @@
 "use client";
 
+import { useSupabase } from "@/components/supabase/supabase-provider";
 import {
   Session,
 } from "@supabase/auth-helpers-nextjs";
-import { useSupabase } from "./components/supabase/supabase-provider";
+
 import { useRouter } from "next/navigation";
 import {  useEffect, useState } from "react";
