+++ src/app/settings/layout.tsx
@@ -1,7 +1,8 @@
 "use client";
 
+import { theme } from "@/utils/theme";
 import { Container, ThemeProvider } from "@mui/material";
-import { theme } from "../utils/theme";
+
 
 export default function SettingsLayout({
   children,
