+++ src/components/dashboard/dashboardHeader.tsx
@@ -1,9 +1,11 @@
 "use client";
 
-import { ThemeProvider, Typography, Card, Box } from "@mui/material";
-import { theme } from "../../utils/theme";
+import { ThemeProvider, Typography, Card } from "@mui/material";
+
 import AddStore from "../utils/addStore";
-import useStore from "../../hooks/useStore";
+
+import { theme } from "@/utils/theme";
+import useStore from "@/hooks/useStore";
 import { useProfileStore } from "@/state/ProfileStore";
 
 export default function DashboardHeader() {
@@ -12,23 +14,26 @@ export default function DashboardHeader() {
   return (
     <>
       <ThemeProvider theme={theme}>
-        {/* make a contianeher here  */}
         {selectId && (
           <Card
-          raised
+            raised
             sx={{
-              boxShadow: 0 ,
-              mt:1,
+              boxShadow: 0,
+              mt: 1,
               display: "flex",
               flexFlow: "row",
               backgroundColor: "white",
             }}
           >
-            <Typography sx={{marginLeft:3}} color="primary.main" variant="h2">
+            <Typography
+              sx={{ marginLeft: 3 }}
+              color="primary.main"
+              variant="h2"
+            >
               Dashboard
             </Typography>
-        
-              <AddStore select_id={selectId} />
+
+            <AddStore select_id={selectId} />
           </Card>
         )}
       </ThemeProvider>
