+++ src/components/groceryStore/groceryStoreHeader.tsx
@@ -4,9 +4,9 @@ import { Box, Button, Card, IconButton, Typography } from "@mui/material";
 import { useParams, useRouter } from "next/navigation";
 import ArrowBackIosNewIcon from "@mui/icons-material/ArrowBackIosNew";
 import GroceryStoreHeaderMenu from "./groceryStoreHeaderMenu";
-import { theme } from "@/app/utils/theme";
+import { theme } from "@/utils/theme";
 import { ThemeProvider } from "@emotion/react";
-import useStore from "@/app/hooks/useStore";
+import useStore from "@/hooks/useStore";
 import { useGroceryStoreStore } from "@/state/GrocerStore";
 
 export default function GroceryStoreHeader() {
