+++ src/components/header/header.tsx
@@ -1,11 +1,13 @@
 "use client";
 
 import { Box, AppBar, Toolbar, IconButton, ThemeProvider } from "@mui/material";
-import { useSupabase } from "../supabase/supabase-provider";
+
 import HomeIcon from "@mui/icons-material/Home";
 import { useParams, useRouter } from "next/navigation";
-import { theme } from "../../utils/theme";
+import { useSupabase } from "../supabase/supabase-provider";
+import { theme } from "@/utils/theme";
 import HeaderMenu from "./headerMenu";
+
 
 export default function Header() {
   const { supabase, session } = useSupabase();
