+++ src/components/header/headerMenu.tsx
@@ -11,9 +11,11 @@ import {
 } from "@mui/material";
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import useStore from "../../hooks/useStore";
-import { useProfileStore } from "@/state/ProfileStore";
+
+
 import { useSupabase } from "../supabase/supabase-provider";
+import useStore from "@/hooks/useStore";
+import { useProfileStore } from "@/state/ProfileStore";
 
 
 export default function HeaderMenu() {
